clean up feed store
