+++ src/renderer/components/Feed/Feeds/GraphFeed.tsx
@@ -1,17 +1,17 @@
-import React, { FC, useCallback, useEffect, useRef, useState } from 'react'
+import { IGraphFeed } from '@/renderer/store/modules/feed'
+import { useInterval } from '@/renderer/utils/hooks/useInterval'
+import { useRosSubscribe } from '@/renderer/utils/hooks/useRosSubscribe'
 import {
   Chart,
   LinearScale,
   LineController,
   LineElement,
-  TimeScale,
   PointElement,
+  TimeScale,
 } from 'chart.js'
-import { IGraphFeed } from '@/renderer/store/modules/feed/@types'
-import _ from 'lodash'
-import { useInterval } from '@/renderer/utils/hooks/useInterval'
-import { useRosSubscribe } from '@/renderer/utils/hooks/useRosSubscribe'
 import 'chartjs-adapter-date-fns'
+import _ from 'lodash'
+import React, { FC, useCallback, useEffect, useRef, useState } from 'react'
 
 Chart.register(
   LineController,
