+++ src/renderer/components/pages/Config/pages/CameraConfig/CameraConfig.tsx
@@ -1,7 +1,7 @@
 import React, { ChangeEvent, FC, useEffect, useState } from 'react'
 import { useDispatch } from 'react-redux'
-import { CameraType } from '@/renderer/store/modules/feed/@types'
-import { feedSlice } from '@/renderer/store/modules/feed/reducer'
+import { CameraType } from '@/renderer/store/modules/feed'
+import { feedSlice } from '@/renderer/store/modules/feed'
 import { LabeledInput } from '@/renderer/components/common/LabeledInput'
 import { Table } from './Table'
 import { Button } from '@/renderer/components/common/Button'
