+++ src/renderer/components/pages/Teleop.tsx
@@ -2,28 +2,7 @@ import React, { FC } from 'react'
 import { Feed } from '@/renderer/components/Feed/Feed'
 import { useControl } from '@/renderer/utils/hooks/useControl'
 import { styled } from '@/renderer/globalStyles/styled'
-import { feed_id } from '@/renderer/store/modules/feed/initialState'
-
-interface PipProps {
-  left?: boolean
-  right?: boolean
-  bottom?: boolean
-  top?: boolean
-}
-
-const StyledPIP = styled.div<PipProps>`
-  position: absolute;
-  width: 384px;
-  height: 216px;
-  z-index: 2;
-
-  border: 1px solid rgba(255, 255, 255, 0.25);
-
-  left: ${(props) => (props.left ? `1px` : undefined)};
-  right: ${(props) => (props.right ? `1px` : undefined)};
-  bottom: ${(props) => (props.bottom ? `1px` : undefined)};
-  top: ${(props) => (props.top ? `1px` : undefined)};
-`
+import { feed_id } from '@/renderer/store/modules/feed'
 
 export const Teleop: FC = () => {
   useControl('flipper')
@@ -42,3 +21,22 @@ export const Teleop: FC = () => {
     </>
   )
 }
+
+const StyledPIP = styled.div<{
+  left?: boolean
+  right?: boolean
+  bottom?: boolean
+  top?: boolean
+}>`
+  position: absolute;
+  width: 384px;
+  height: 216px;
+  z-index: 2;
+
+  border: 1px solid rgba(255, 255, 255, 0.25);
+
+  left: ${(props) => (props.left ? `1px` : undefined)};
+  right: ${(props) => (props.right ? `1px` : undefined)};
+  bottom: ${(props) => (props.bottom ? `1px` : undefined)};
+  top: ${(props) => (props.top ? `1px` : undefined)};
+`
