+++ src/renderer/store/localStorage.ts
@@ -1,6 +1,6 @@
 import { GlobalState } from '@/renderer/store/store'
 import shortid from 'shortid'
-import { initialState as feedState } from '@/renderer/store/modules/feed/initialState'
+import { initialState as feedState } from '@/renderer/store/modules/feed'
 import { initialState as rosState } from '@/renderer/store/modules/ros'
 
 export const defaultState: GlobalState = {
